Pull request cherry-pick helper
This is a modified version of old cherry-pick-upstream.py which supports
continuing a pull request merge which failed by a conflict.

Usage:

    prx.py merge [PR ID]

This will merge a pull request with an ID, e.g. 10460 or so. The merge
will stop when a cherry-pick fails with conflict, and the state will be
stored. The user should manually fix the conflict, commit the chagnes
and use prx.py continue to continue the process.

The merge command will fail if a current prx.py merge is undergoing.

NOTE: This merge will have its own internal state, and is different from
the standard git merge.

    prx.py abort

Abort the current merge.

    prx.py continue

Continue the current merge, assuming the conflicting commit is resolved
and can be skipped.

    prx.py status

Report the current status of merge.